asdf

+++ hdf5_functions.py
@@ -338,7 +338,7 @@ def dir2phasesol(mtf, ms='', directions=[]):
     source_table.append(source_soltab.items())  # from dictionary to list
     antenna_table = table.obj._f_get_child('antenna')
     antenna_table.append(antenna_soltab.items())  # from dictionary to list
-    scoop = {'POINTING': np.array([directions.ra.rad, directions.dec.rad])}
+    scoop = {'POINTING': np.array([directions.ra.rad, directions.dec.rad], dtype='float32')}
     print('DIRS', scoop)
     print('sols', source_soltab)
     print('solkeys', source_soltab.keys())
